remove feature flag, substitute real one for callout feature
