+++ apps/web/src/app/admin-console/organizations/members/services/invite-link-callout/invite-link-callout.service.ts
@@ -79,7 +79,7 @@ export class InviteLinkCalloutService {
       return;
     }
 
-    if (!(await this.configService.getFeatureFlag(FeatureFlag.GenerateInviteLink))) {
+    if (!(await this.configService.getFeatureFlag(FeatureFlag.InviteLinkNotification))) {
       return;
     }
