+++ apps/web/src/app/auth/core/services/login/web-login-component.service.ts
@@ -11,14 +11,11 @@ import {
   PasswordPolicies,
 } from "@bitwarden/auth/angular";
 import { InternalPolicyService } from "@bitwarden/common/admin-console/abstractions/policy/policy.service.abstraction";
-import { AccountService } from "@bitwarden/common/auth/abstractions/account.service";
 import { SsoLoginServiceAbstraction } from "@bitwarden/common/auth/abstractions/sso-login.service.abstraction";
 import {
   OrgInviteKind,
   OrganizationInviteService,
 } from "@bitwarden/common/auth/organization-invite";
-import { FeatureFlag } from "@bitwarden/common/enums/feature-flag.enum";
-import { ConfigService } from "@bitwarden/common/platform/abstractions/config/config.service";
 import { EnvironmentService } from "@bitwarden/common/platform/abstractions/environment.service";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { LogService } from "@bitwarden/common/platform/abstractions/log.service";
@@ -62,8 +59,6 @@ export class WebLoginComponentService
     platformUtilsService: PlatformUtilsService,
     ssoLoginService: SsoLoginServiceAbstraction,
     private router: Router,
-    private accountService: AccountService,
-    private configService: ConfigService,
     private toastService: ToastService,
     private i18nService: I18nService,
   ) {
@@ -251,15 +246,6 @@ export class WebLoginComponentService
         this.logService.error(
           `WebLoginComponentService.getOrgPoliciesFromOrgInvite: Email mismatch. Expected: ${orgInvite.email}, Received: ${email}`,
         );
-        return undefined;
-      }
-    } else {
-      // Defense in depth: stale flag-on state may persist into a flag-off session.
-      // Treat as no invite when disabled.
-      // TODO: clean up when FeatureFlag.GenerateInviteLink is removed — drop this
-      // guard clause; the `else` branch can be removed entirely and the outer `if`
-      // collapsed since the direct-invite branch above is the only remaining case.
-      if (!(await this.configService.getFeatureFlag(FeatureFlag.GenerateInviteLink))) {
         return undefined;
       }
     }
