+++ apps/web/src/app/auth/core/services/registration/web-registration-finish.service.spec.ts
@@ -135,9 +135,6 @@ describe("WebRegistrationFinishService", () => {
       legacyCompatKeyService.makeUserKey.mockResolvedValue([legacyUserKey, userKeyEncString]);
       legacyCompatKeyService.makeKeyPair.mockResolvedValue(userKeyPair);
       accountApiService.registerFinish.mockResolvedValue();
-      configService.getFeatureFlag.mockImplementation((flag) =>
-        Promise.resolve(flag === FeatureFlag.GenerateInviteLink),
-      );
 
       masterPasswordService.makeMasterPasswordAuthenticationData.mockResolvedValue({
         salt,
@@ -189,10 +186,7 @@ describe("WebRegistrationFinishService", () => {
     label: "SDK flow",
     setupMocks: () => {
       configService.getFeatureFlag.mockImplementation((flag) =>
-        Promise.resolve(
-          flag === FeatureFlag.EnableAccountEncryptionV2UserPasswordRegistration ||
-            flag === FeatureFlag.GenerateInviteLink,
-        ),
+        Promise.resolve(flag === FeatureFlag.EnableAccountEncryptionV2UserPasswordRegistration),
       );
 
       postKeysForUserPasswordRegistration = jest.fn().mockResolvedValue(undefined);
@@ -311,30 +305,6 @@ describe("WebRegistrationFinishService", () => {
 
     it("does not populate open-org-invite fields when a direct org invite is stashed", async () => {
       organizationInviteService.getOrganizationInvite.mockResolvedValue(variant.directOrgInvite);
-
-      await service.finishRegistration(email, passwordInputResult);
-
-      const request = variant.getRequest();
-      expect(request[variant.fields.openOrgInvite]).toBeUndefined();
-    });
-
-    it("does not populate open-org-invite fields when GenerateInviteLink flag is off, even if an open invite is stashed", async () => {
-      // Defense in depth: an open invite stashed while the flag was on must not leak
-      // into a flag-off session.
-      configService.getFeatureFlag.mockImplementation((flag) =>
-        Promise.resolve(
-          flag === FeatureFlag.EnableAccountEncryptionV2UserPasswordRegistration &&
-            variant.label === "SDK flow",
-        ),
-      );
-      organizationInviteService.getOrganizationInvite.mockResolvedValue(
-        new OpenOrganizationInvite({
-          organizationId: "00000000-0000-0000-0000-000000000010",
-          inviteLinkCode: "00000000-0000-0000-0000-000000000011",
-          inviteKey: "link-key",
-          organizationName: "openOrgName",
-        }),
-      );
 
       await service.finishRegistration(email, passwordInputResult);
