+++ apps/web/src/app/auth/core/services/registration/web-registration-finish.service.ts
@@ -12,7 +12,6 @@ import {
   OrganizationInviteService,
   OrgInviteKind,
 } from "@bitwarden/common/auth/organization-invite";
-import { FeatureFlag } from "@bitwarden/common/enums/feature-flag.enum";
 import { MasterPasswordServiceAbstraction } from "@bitwarden/common/key-management/master-password/abstractions/master-password.service.abstraction";
 import { ConfigService } from "@bitwarden/common/platform/abstractions/config/config.service";
 import { asUuid, SdkService } from "@bitwarden/common/platform/abstractions/sdk/sdk.service";
@@ -88,13 +87,7 @@ export class WebRegistrationFinishService
         orgInvite.organizationUserId,
       );
       registerRequest.org_invite_token = orgInvite.token;
-    } else if (
-      orgInvite?.kind === OrgInviteKind.Open &&
-      // Defense in depth: stale flag-on state may persist into a flag-off session.
-      // TODO: clean up when FeatureFlag.GenerateInviteLink is removed — drop this
-      // guard clause.
-      (await this.configService.getFeatureFlag(FeatureFlag.GenerateInviteLink))
-    ) {
+    } else if (orgInvite?.kind === OrgInviteKind.Open) {
       registerRequest.open_org_invite = {
         organization_id: asUuid<SdkOrganizationId>(orgInvite.organizationId),
         code: orgInvite.inviteLinkCode,
@@ -177,13 +170,7 @@ export class WebRegistrationFinishService
     if (orgInvite?.kind === OrgInviteKind.Direct) {
       registerRequest.organizationUserId = orgInvite.organizationUserId;
       registerRequest.orgInviteToken = orgInvite.token;
-    } else if (
-      orgInvite?.kind === OrgInviteKind.Open &&
-      // Defense in depth: stale flag-on state may persist into a flag-off session.
-      // TODO: clean up when FeatureFlag.GenerateInviteLink is removed — drop this
-      // guard clause.
-      (await this.configService.getFeatureFlag(FeatureFlag.GenerateInviteLink))
-    ) {
+    } else if (orgInvite?.kind === OrgInviteKind.Open) {
       registerRequest.openOrgInvite = new OpenOrgInviteRequest(
         orgInvite.organizationId,
         orgInvite.inviteLinkCode,
