+++ bitwarden_license/bit-web/src/app/admin-console/organizations/organizations-routing.module.ts
@@ -1,15 +1,11 @@
-import { inject, NgModule } from "@angular/core";
+import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
-import { map } from "rxjs";
 
 import { authGuard } from "@bitwarden/angular/auth/guards";
-import { componentRouteSwap } from "@bitwarden/angular/utils/component-route-swap";
 import {
   canAccessAccessIntelligence,
   canAccessSettingsTab,
 } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
-import { FeatureFlag } from "@bitwarden/common/enums/feature-flag.enum";
-import { ConfigService } from "@bitwarden/common/platform/abstractions/config/config.service";
 import { safeProvider } from "@bitwarden/ui-common";
 import { isEnterpriseOrgGuard } from "@bitwarden/web-vault/app/admin-console/organizations/guards/is-enterprise-org.guard";
 import { organizationPermissionsGuard } from "@bitwarden/web-vault/app/admin-console/organizations/guards/org-permissions.guard";
@@ -21,7 +17,6 @@ import { BillingConstraintService } from "@bitwarden/web-vault/app/billing/membe
 import { SsoManageComponent } from "../../auth/sso/sso-manage.component";
 
 import { DomainVerificationComponent } from "./manage/domain-verification/domain-verification.component";
-import { ScimV2Component } from "./manage/scim-v2.component";
 import { ScimComponent } from "./manage/scim.component";
 
 const routes: Routes = [
@@ -50,24 +45,12 @@ const routes: Routes = [
               titleId: "singleSignOn",
             },
           },
-          ...componentRouteSwap(
-            ScimComponent,
-            ScimV2Component,
-            () =>
-              inject(ConfigService)
-                .getFeatureFlag$(FeatureFlag.GenerateInviteLink)
-                .pipe(map((v) => v === true)),
-            {
-              path: "scim",
-              canActivate: [organizationPermissionsGuard((org) => org.canManageScim)],
-              data: { titleId: "scim" },
-            },
-            {
-              path: "scim",
-              canActivate: [organizationPermissionsGuard((org) => org.canManageScim)],
-              data: { titleId: "scimV2" },
-            },
-          ),
+          {
+            path: "scim",
+            component: ScimComponent,
+            canActivate: [organizationPermissionsGuard((org) => org.canManageScim)],
+            data: { titleId: "scimV2" },
+          },
           {
             path: "device-approvals",
             loadComponent: () =>
