+++ bitwarden_license/bit-web/src/app/admin-console/organizations/organizations.module.ts
@@ -8,17 +8,11 @@ import { SsoManageComponent } from "../../auth/sso/sso-manage.component";
 
 import { DomainAddEditDialogComponent } from "./manage/domain-verification/domain-add-edit-dialog/domain-add-edit-dialog.component";
 import { DomainVerificationComponent } from "./manage/domain-verification/domain-verification.component";
-import { ScimV2Component } from "./manage/scim-v2.component";
 import { ScimComponent } from "./manage/scim.component";
 import { OrganizationsRoutingModule } from "./organizations-routing.module";
 
 @NgModule({
-  imports: [SharedModule, OrganizationsRoutingModule, HeaderModule, ScimV2Component, IconModule],
-  declarations: [
-    SsoManageComponent,
-    ScimComponent,
-    DomainVerificationComponent,
-    DomainAddEditDialogComponent,
-  ],
+  imports: [SharedModule, OrganizationsRoutingModule, HeaderModule, ScimComponent, IconModule],
+  declarations: [SsoManageComponent, DomainVerificationComponent, DomainAddEditDialogComponent],
 })
 export class OrganizationsModule {}
