+++ libs/auth/src/angular/login/login.component.ts
@@ -29,11 +29,9 @@ import { AuthResult } from "@bitwarden/common/auth/models/domain/auth-result";
 import { OrganizationInviteService } from "@bitwarden/common/auth/organization-invite";
 import { PasswordPreloginService } from "@bitwarden/common/auth/password-prelogin";
 import { ClientType, HttpStatusCode } from "@bitwarden/common/enums";
-import { FeatureFlag } from "@bitwarden/common/enums/feature-flag.enum";
 import { ErrorResponse } from "@bitwarden/common/models/response/error.response";
 import { AppIdService } from "@bitwarden/common/platform/abstractions/app-id.service";
 import { BroadcasterService } from "@bitwarden/common/platform/abstractions/broadcaster.service";
-import { ConfigService } from "@bitwarden/common/platform/abstractions/config/config.service";
 import { EnvironmentService } from "@bitwarden/common/platform/abstractions/environment.service";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { LogService } from "@bitwarden/common/platform/abstractions/log.service";
@@ -147,7 +145,6 @@ export class LoginComponent implements OnInit, OnDestroy {
     private logService: LogService,
     private validationService: ValidationService,
     private loginSuccessHandlerService: LoginSuccessHandlerService,
-    private configService: ConfigService,
     private ssoLoginService: SsoLoginServiceAbstraction,
     private environmentService: EnvironmentService,
     private passwordPreloginService: PasswordPreloginService,
@@ -657,13 +654,6 @@ export class LoginComponent implements OnInit, OnDestroy {
     if (invite == null) {
       return true;
     }
-    // Defense in depth: stale flag-on state may persist into a flag-off session.
-    // Skip the domain check when disabled.
-    // TODO: clean up when FeatureFlag.GenerateInviteLink is removed — drop this
-    // guard clause.
-    if (!(await this.configService.getFeatureFlag(FeatureFlag.GenerateInviteLink))) {
-      return true;
-    }
     const result = await this.organizationInviteService.validateOpenOrgInviteEmailDomain(
       invite.organizationId,
       invite.inviteLinkCode,
@@ -690,23 +680,12 @@ export class LoginComponent implements OnInit, OnDestroy {
   }
 
   /**
-   * Returns the open-org-invite if one is in state and the feature is enabled; otherwise
-   * `null`. Centralizes the "should we apply open-org-invite chrome here?" predicate so
-   * the kind + flag guard isn't restated at every override site.
-   *
-   * Defense in depth: stale flag-on state may persist into a flag-off session.
+   * Returns the open-org-invite if one is in state; otherwise `null`. Centralizes the
+   * "should we apply open-org-invite chrome here?" predicate so the kind guard isn't
+   * restated at every override site.
    */
   private async getActiveOpenOrgInvite(): Promise<{ organizationName: string } | null> {
-    const invite = await this.organizationInviteService.getOpenOrgInvite();
-    if (invite == null) {
-      return null;
-    }
-    // TODO: clean up when FeatureFlag.GenerateInviteLink is removed — drop this
-    // guard clause.
-    if (!(await this.configService.getFeatureFlag(FeatureFlag.GenerateInviteLink))) {
-      return null;
-    }
-    return invite;
+    return await this.organizationInviteService.getOpenOrgInvite();
   }
 
   /**
