+++ libs/auth/src/angular/registration/registration-start/registration-start.component.ts
@@ -12,8 +12,6 @@ import { AccountApiService } from "@bitwarden/common/auth/abstractions/account-a
 import { RegisterSendVerificationEmailRequest } from "@bitwarden/common/auth/models/request/registration/register-send-verification-email.request";
 import { RegisterStartOpenOrgInviteRequest } from "@bitwarden/common/auth/models/request/registration/register-start-open-org-invite.request";
 import { OrganizationInviteService } from "@bitwarden/common/auth/organization-invite";
-import { FeatureFlag } from "@bitwarden/common/enums/feature-flag.enum";
-import { ConfigService } from "@bitwarden/common/platform/abstractions/config/config.service";
 import { RegionConfig, Region } from "@bitwarden/common/platform/abstractions/environment.service";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { PlatformUtilsService } from "@bitwarden/common/platform/abstractions/platform-utils.service";
@@ -110,7 +108,6 @@ export class RegistrationStartComponent implements OnInit, OnDestroy {
     private anonLayoutWrapperDataService: AnonLayoutWrapperDataService,
     private organizationInviteService: OrganizationInviteService,
     private i18nService: I18nService,
-    private configService: ConfigService,
     private validationService: ValidationService,
   ) {
     this.isSelfHost = platformUtilsService.isSelfHost();
@@ -139,17 +136,10 @@ export class RegistrationStartComponent implements OnInit, OnDestroy {
    * "Join <organizationName>" so users see they're accepting an invite rather than
    * generic "Create account" chrome. No icon override — the existing route-data icon
    * (RegistrationUserAddIcon) is reused.
-   *
-   * Defense in depth: stale flag-on state may persist into a flag-off session.
    */
   private async applyOpenOrgInviteTitleOverride(): Promise<void> {
     const invite = await this.organizationInviteService.getOpenOrgInvite();
     if (invite == null) {
-      return;
-    }
-    // TODO: clean up when FeatureFlag.GenerateInviteLink is removed — drop this
-    // guard clause.
-    if (!(await this.configService.getFeatureFlag(FeatureFlag.GenerateInviteLink))) {
       return;
     }
     this.anonLayoutWrapperDataService.setAnonLayoutWrapperData({
@@ -193,24 +183,13 @@ export class RegistrationStartComponent implements OnInit, OnDestroy {
     // The app expects null for name and not empty string.
     const sanitizedName = this.name.value === "" ? null : this.name.value;
 
-    // TODO: clean up when FeatureFlag.GenerateInviteLink is removed — collapse to the
-    // on-flag branch.
-    let request: RegisterSendVerificationEmailRequest;
-    if (await this.configService.getFeatureFlag(FeatureFlag.GenerateInviteLink)) {
-      const openOrgInviteRequest = await this.buildOpenOrgInviteRequestIfPresent(this.email.value);
-      request = new RegisterSendVerificationEmailRequest(
-        this.email.value,
-        sanitizedName,
-        this.receiveMarketingEmails.value,
-        openOrgInviteRequest,
-      );
-    } else {
-      request = new RegisterSendVerificationEmailRequest(
-        this.email.value,
-        sanitizedName,
-        this.receiveMarketingEmails.value,
-      );
-    }
+    const openOrgInviteRequest = await this.buildOpenOrgInviteRequestIfPresent(this.email.value);
+    const request = new RegisterSendVerificationEmailRequest(
+      this.email.value,
+      sanitizedName,
+      this.receiveMarketingEmails.value,
+      openOrgInviteRequest,
+    );
 
     const result = await this.accountApiService.registerSendVerificationEmail(request);
 
@@ -312,13 +291,6 @@ export class RegistrationStartComponent implements OnInit, OnDestroy {
   private async openOrgInviteDomainAllowed(email: string): Promise<boolean> {
     const invite = await this.organizationInviteService.getOpenOrgInvite();
     if (invite == null) {
-      return true;
-    }
-    // Defense in depth: stale flag-on state may persist into a flag-off session.
-    // Skip the domain check when disabled.
-    // TODO: clean up when FeatureFlag.GenerateInviteLink is removed — drop this
-    // guard clause.
-    if (!(await this.configService.getFeatureFlag(FeatureFlag.GenerateInviteLink))) {
       return true;
     }
     const result = await this.organizationInviteService.validateOpenOrgInviteEmailDomain(
